Fix control message panic on close
