+++ pkg/client/rest/types.go
@@ -5,10 +5,19 @@ package rest
 
 import (
 	"time"
+
+	simple "github.com/dvonthenen/symbl-go-sdk/pkg/client/simple"
 )
 
 // AccessToken represents a Symbl platform bearer access token with expiry information.
 type AccessToken struct {
 	AccessToken string
 	ExpiresOn   time.Time
+}
+
+// Client which extends basic client to support REST
+type Client struct {
+	*simple.Client
+
+	auth *AccessToken
 }
